Extend dijkstra to allow for monotone A* heuristics
The new template parameter ComparatorT allows to use another comparator
in the priority queue of the dijkstra algorithm, that can be used for
comparators that use a A* type heuristic function, e.g., current path
length + norm(point - target_point).

This is not a full A* implementation, as it requires the distance
function in the comparator to be monotone, while a complete A*
implementation allows for non-monotone heuristics by visiting vertices
multiple times.