Added readOBJ / writeOBJ methods that support arbitrary scalar types
The OpenMesh functions for reading and writing meshes use float as data
type instead of the scalar defined in the mesh type traits. This
functions use the C++ auto keyword to automatically use the right scalar
for the mesh type.
writeOBJ has an parameter for the wanted precision floating point
precision.